add popup on logout

+++ src/Components/Home_Page/Navbar.jsx
@@ -306,10 +306,13 @@ export default function Navbar(props) {
               <div className="hr"></div>
               <div
                 onClick={() => {
-                  localStorage.setItem("isLogged", JSON.stringify(false));
-                  localStorage.setItem("authTok", JSON.stringify(""));
-                  Navigation("/login");
-                }}
+    const shouldLogout = window.confirm("Are you sure you want to logout?");
+    if (shouldLogout) {
+      localStorage.setItem("isLogged", JSON.stringify(false));
+      localStorage.setItem("authTok", JSON.stringify(""));
+      Navigation("/login");
+    }
+  }}
               >
                 Logout
               </div>
